+++ README.md
@@ -40,7 +40,7 @@
 
     - [x] RESTAURANT_ITEMS를 직관적인 네이밍으로 수정
     - [x] link나 description에 대한 예외 처리 및 테스트 추가
-    - [ ] 컴포넌트를 통일된 형식으로 수정 + 상태 관리 레이어 추가
+    - [x] 컴포넌트를 통일된 형식으로 수정 + 상태 관리 레이어 추가
 
 <br/>
 
@@ -50,7 +50,7 @@
 
 ## ⚙️ 기능 구현
 
-    - [ ] 음식점의 상세 정보를 확인하고, 자주 가는 음식점으로 지정할 수 있는 기능을 추가한다.
+    음식점의 상세 정보를 확인하고, 자주 가는 음식점으로 지정할 수 있는 기능을 추가한다.
     - [ ] 음식점 목록을 확인할 수 있다.
       - [ ] 카테고리별로 필터링해서 확인할 수 있다.
       - [ ] 이름순/거리순으로 정렬해서 확인할 수 있다.
