+++ src/components/RestaurantEnrollModal.js
@@ -1,10 +1,27 @@
 import createElement from '../utils/createElement.js';
-import Modal from './Modal.js';
-import { createRestaurantEnrollForm } from './RestaurantEnrollForm.js';
-import { resetInput } from './RestaurantEnrollForm.js';
+import RestaurantValidator from '../validators/RestaurantValidator.js';
+import Modal from './common/Modal.js';
+import createRestaurantEnrollForm from './RestaurantEnrollForm.js';
+import { updateRestaurantList } from './RestaurantList.js';
+
+function handleClose() {
+  document.querySelector('select#category').value = '';
+  document.querySelector('input#name').value = '';
+  document.querySelector('select#distance').value = '';
+  document.querySelector('textarea#description').value = '';
+  document.querySelector('input#link').value = '';
+}
 
 function createRestaurantEnrollModal() {
-  const modal = new Modal();
+  const modal = new Modal(handleClose);
+
+  const restaurantInput = {
+    name: null,
+    category: null,
+    distance: null,
+    description: null,
+    link: null,
+  };
 
   const $modalTitle = createElement({
     tag: 'h2',
@@ -12,7 +29,17 @@ function createRestaurantEnrollModal() {
     textContent: '새로운 음식점',
   });
 
-  const $enrollForm = createRestaurantEnrollForm(() => modal.toggle());
+  const $enrollForm = createRestaurantEnrollForm(
+    restaurantInput,
+    (event) => {
+      const isValidate = RestaurantValidator.validate(restaurantInput);
+      if (!isValidate) return;
+
+      updateRestaurantList(restaurantInput);
+      modal.toggle();
+    },
+    () => modal.toggle()
+  );
 
   const fragment = new DocumentFragment();
   fragment.append($modalTitle, $enrollForm);
