+++ src/components/common/InputBox.js
@@ -1,4 +1,4 @@
-import createElement from '../utils/createElement.js';
+import createElement from '../../utils/createElement.js';
 
 function createInputBox({ label, isRequired, type, helpText = '', onChange }) {
   const inputBoxDiv = createElement({
